requirements.txt
tensorflow
streamlit
pyyaml
pillow
numpy